refactor(webapp): [dev] Small changes to webapp backend code structure.
