+++ suricata/rules/suricata.rules
@@ -1,12 +1,12 @@
 
-alert ip any any -> any any (msg: "A CINI flag (CCIT) was sent to client"; flow:to_client; content: "="; pcre: "/^[A-Z0-9]{31}=$/, flow:match"; distance: -32; content:!"AAAAA="; distance: -6; metadata: tag FLAG OUT, color danger; flowint: tag_FLAG_OUT,+,1; sid: 1;)
-alert ip any any -> any any (msg: "A CINI flag (CCIT) was sent to client"; flow:to_client; file.data; content: "="; pcre: "/^[A-Z0-9]{31}=$/, flow:match"; distance: -32; content:!"AAAAA="; distance: -6; metadata: tag FLAG OUT, color danger; flowint: tag_FLAG_OUT,+,1; sid: 2;)
-alert ip any any -> any any (msg: "A CINI flag (CCIT) was sent to client (hex)"; flow:to_client; content: "3d"; pcre: "/^[0-9a-f]{62}3d$/, flow:match"; distance: -64; metadata: tag FLAG OUT HEX, color danger; flowint: tag_FLAG_OUT_HEX,+,1; sid: 3;)
-alert ip any any -> any any (msg: "A CINI flag (CCIT) was sent to client (hex)"; flow:to_client; content: "3D"; pcre: "/^[0-9A-F]{62}3D$/, flow:match"; distance: -64; metadata: tag FLAG OUT HEX, color danger; flowint: tag_FLAG_OUT_HEX,+,1; sid: 4;)
-alert ip any any -> any any (msg: "A CINI flag (CCIT) was sent to client (base64)"; flow:to_client; content: "0="; pcre: "/^[A-Za-z0-9+\/]{42}0=$/, flow:match"; distance: -44; metadata: tag FLAG OUT B64, color danger; flowint: tag_FLAG_OUT_B64,+,1; sid: 5;)
-alert ip any any -> any any (msg: "A CINI flag (CCIT) was sent to client (URL encoded)"; flow:to_client; content: "%3d"; pcre: "/^[A-Za-z0-9]{31}%3d$/, flow:match"; distance: -34; metadata: tag FLAG OUT B64, color danger; flowint: tag_FLAG_OUT_URL,+,1; sid: 6;)
-alert ip any any -> any any (msg: "A CINI flag (CCIT) was sent to client (URL encoded)"; flow:to_client; content: "%3D"; pcre: "/^[A-Za-z0-9]{31}%3D$/, flow:match"; distance: -34; metadata: tag FLAG OUT B64, color danger; flowint: tag_FLAG_OUT_URL,+,1; sid: 7;)
-alert ip any any -> any any (msg: "A CINI flag (CCIT) was placed in our services (probably by checkers)"; flow:to_server; content: "="; pcre: "/^[A-Z0-9]{31}=$/, flow:match"; distance: -32; content:!"AAAAA="; distance: -6; metadata: tag FLAG IN, color success; flowint: tag_FLAG_IN,+,1; sid: 8;)
+alert ip any any -> any any (msg: "A saarCTF flag was sent to client"; flow:to_client; content: "SAAR"; pcre: "/(SAAR\{[A-Za-z0-9-_]{32}\})/, flow:match"; distance: -4; metadata: tag FLAG OUT, color danger; flowint: tag_FLAG_OUT,+,1; sid: 1;)
+alert ip any any -> any any (msg: "A saarCTF flag was sent to client"; flow:to_client; file.data; content: "SAAR{"; pcre: "/(SAAR\{[A-Za-z0-9-_]{32}\})/, flow:match"; distance: -5; metadata: tag FLAG OUT, color danger; flowint: tag_FLAG_OUT,+,1; sid: 2;)
+alert ip any any -> any any (msg: "A saarCTF flag was sent to client (URL encoded)"; flow:to_client; content: "SAAR%7B"; pcre: "/(SAAR%7B[A-Za-z0-9-_]{32}%7D)/, flow:match"; distance: -7; metadata: tag FLAG OUT, color danger; flowint: tag_FLAG_OUT,+,1; sid: 3;)
+alert ip any any -> any any (msg: "A saarCTF flag was sent to client (base64)"; flow:to_client; content: "U0FBUnt"; metadata: tag FLAG OUT B64, color danger; flowint: tag_FLAG_OUT_B64,+,1; sid: 4;)
+alert ip any any -> any any (msg: "A saarCTF flag was sent to client (base64)"; flow:to_client; content: "NBQVJ7"; pcre: "/(NBQVJ7[A-Za-z\d]{43}9)/, flow:match"; distance: -6; metadata: tag FLAG OUT B64, color danger; flowint: tag_FLAG_OUT_B64,+,1; sid: 5;)
+alert ip any any -> any any (msg: "A saarCTF flag was sent to client (base64)"; flow:to_client; content: "TQUFSe"; pcre: "/(TQUFSe[A-Za-z\d]{43}f)/, flow:match"; distance: -6; metadata: tag FLAG OUT B64, color danger; flowint: tag_FLAG_OUT_B64,+,1; sid: 6;)
+alert ip any any -> any any (msg: "A saarCTF flag was placed in our services (probably by checkers)"; flow:to_server; content: "SAAR"; pcre: "/(SAAR\{[A-Za-z0-9-_]{32}\})/, flow:match"; distance: -4; metadata: tag FLAG IN, color success; flowint: tag_FLAG_IN,+,1; sid: 7;)
+alert ip any any -> any any (msg: "A saarCTF flag was placed in our services (probably by checkers, URL encoded)"; flow:to_server; content: "SAAR%7B"; pcre: "/(SAAR%7B[A-Za-z0-9-_]{32}%7D)/, flow:match"; distance: -7; metadata: tag FLAG IN, color success; flowint: tag_FLAG_IN,+,1; sid: 8;)
 
 # Copyright (C) 2023-2024  ANSSI
 # SPDX-License-Identifier: GPL-2.0-or-later
