+++ webapp/backend/src/config.rs
@@ -1,10 +1,8 @@
+use std::sync::Mutex;
+use actix_web::{delete, get, post, web, Responder, Result};
 use std::{collections::HashMap, fs::File, io::{Read, Write}};
 use chrono::Duration;
 use serde::{Deserialize, Serialize};
-
-mod routes;
-
-pub use routes::init_routes;
 
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -88,4 +86,73 @@ impl IpPort {
     pub fn to_string(&self) -> String {
         self.ip.clone() + ":" + &self.port.to_string()
     }
+}
+
+
+pub fn init_routes(cfg: &mut web::ServiceConfig) {
+    cfg.service(read);
+    cfg.service(update);
+    cfg.service(read_services);
+    cfg.service(update_services);
+    cfg.service(delete_services);
+}
+
+#[get("/api/config")]
+async fn read(ctf_config: web::Data<Mutex<CtfConfig>>) -> Result<impl Responder> {
+    let ctf_config = ctf_config.lock().unwrap();
+
+    Ok(web::Json(ctf_config.to_owned()))
+}
+
+#[post("/api/config")]
+async fn update(new_ctf_config: web::Json<NewCtfConfig>, ctf_config: web::Data<Mutex<CtfConfig>>) -> Result<impl Responder> {
+    let mut ctf_config = ctf_config.lock().unwrap();
+    ctf_config.start_date = new_ctf_config.start_date.clone();
+    ctf_config.end_date = new_ctf_config.end_date.clone();
+    ctf_config.tick_length = new_ctf_config.tick_length;
+    
+    if let Err(_e) = ctf_config.save("./ctf_config.json") {
+        println!("Error: Can't save JSON to `ctf_config.json`.");
+    }
+
+    Ok(web::Json(ctf_config.to_owned()))
+}
+
+#[get("/api/config/services")]
+async fn read_services(ctf_config: web::Data<Mutex<CtfConfig>>) -> Result<impl Responder> {
+    let ctf_config = ctf_config.lock().unwrap();
+
+    Ok(web::Json(ctf_config.services.to_owned())) 
+}
+
+#[post("/api/config/services")]
+async fn update_services(new_service: web::Json<NewService>, ctf_config: web::Data<Mutex<CtfConfig>>) -> Result<impl Responder> {
+    let mut ctf_config = ctf_config.lock().unwrap();
+    ctf_config.services.insert(new_service.name.clone(), Service {
+        ipports: new_service.ipports.clone(),
+        color: new_service.color.clone()
+    });
+
+    if let Err(_e) = ctf_config.save("./ctf_config.json") {
+        println!("Error: Can't save JSON to `ctf_config.json`.");
+    }
+
+    Ok(web::Json(ctf_config.to_owned())) 
+}
+
+#[derive(Deserialize)]
+struct DeleteService {
+    name: String
+}
+
+#[delete("/api/config/services")]
+async fn delete_services(query: web::Query<DeleteService>, ctf_config: web::Data<Mutex<CtfConfig>>) -> Result<impl Responder> {
+    let mut ctf_config = ctf_config.lock().unwrap();
+    ctf_config.services.remove(&query.name);
+
+    if let Err(_e) = ctf_config.save("./ctf_config.json") {
+        println!("Error: Can't save JSON to `ctf_config.json`.");
+    }
+    
+    Ok(web::Json(ctf_config.services.to_owned())) 
 }
